Better handling of default values

+++ ibuilder/tests/compile/default.rs
@@ -3,7 +3,7 @@ use ibuilder::*;
 #[derive(IBuilder)]
 pub struct Foo {
     #[ibuilder(default = 42)]
-    bar: i32,
+    bar: u8,
 }
 
 fn main() {
